move RuleCodeAndBody to grouped, where it's used
