+++ crates/ruff_linter/src/message/grouped.rs
@@ -10,7 +10,6 @@ use ruff_source_file::OneIndexed;
 
 use crate::fs::relativize_path;
 use crate::message::diff::calculate_print_width;
-use crate::message::text::RuleCodeAndBody;
 use crate::message::{Emitter, EmitterContext, MessageWithLocation, group_diagnostics_by_filename};
 use crate::settings::types::UnsafeFixes;
 
@@ -141,6 +140,49 @@ impl Display for DisplayGroupedMessage<'_> {
         )?;
 
         Ok(())
+    }
+}
+
+pub(super) struct RuleCodeAndBody<'a> {
+    pub(crate) message: &'a Diagnostic,
+    pub(crate) show_fix_status: bool,
+    pub(crate) unsafe_fixes: UnsafeFixes,
+}
+
+impl Display for RuleCodeAndBody<'_> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        if self.show_fix_status {
+            if let Some(fix) = self.message.fix() {
+                // Do not display an indicator for inapplicable fixes
+                if fix.applies(self.unsafe_fixes.required_applicability()) {
+                    if let Some(code) = self.message.secondary_code() {
+                        write!(f, "{} ", code.red().bold())?;
+                    }
+                    return write!(
+                        f,
+                        "{fix}{body}",
+                        fix = format_args!("[{}] ", "*".cyan()),
+                        body = self.message.body(),
+                    );
+                }
+            }
+        }
+
+        if let Some(code) = self.message.secondary_code() {
+            write!(
+                f,
+                "{code} {body}",
+                code = code.red().bold(),
+                body = self.message.body(),
+            )
+        } else {
+            write!(
+                f,
+                "{code}: {body}",
+                code = self.message.id().as_str().red().bold(),
+                body = self.message.body(),
+            )
+        }
     }
 }
